# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabricGhar API specification — shared shop data scoped by user email
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EnsureAccount {
  email: string;
}

export interface AuthResult {
  ok: boolean;
  error?: string;
}

export interface Product {
  id: string;
  name: string;
  cat: string;
  buy: number;
  sell: number;
  stock: number;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  cat: string;
  buy: number;
  sell: number;
  stock: number;
}

export interface Sale {
  id: string;
  /** ISO date string */
  date: string;
  pid: string;
  pname: string;
  qty: number;
  price: number;
  buy: number;
  disc: number;
  total: number;
  profit: number;
  note?: string;
}

export interface SaleInput {
  pid: string;
  /** @minimum 1 */
  qty: number;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  disc: number;
  note?: string;
}

export interface Expense {
  id: string;
  /** ISO date string */
  date: string;
  cat: string;
  amount: number;
  note?: string;
}

export interface ExpenseInput {
  /** @minLength 1 */
  cat: string;
  /** @minimum 0 */
  amount: number;
  note?: string;
  /** Optional ISO date string (defaults to now) */
  date?: string;
}

export interface DashboardSummary {
  todayRevenue: number;
  todayProfit: number;
  totalRevenue: number;
  totalProfit: number;
  productCount: number;
  lowStockCount: number;
  todaySales: Sale[];
}

export interface MonthlyReportEntry {
  /** YYYY-MM */
  month: string;
  revenue: number;
  profit: number;
  count: number;
}

export interface ProductReportEntry {
  name: string;
  revenue: number;
  profit: number;
  qty: number;
}

export interface ReportData {
  totalRevenue: number;
  totalCost: number;
  totalProfit: number;
  profitMargin: number;
  byMonth: MonthlyReportEntry[];
  topProducts: ProductReportEntry[];
}

export interface BackupSnapshot {
  exportedAt?: string;
  products: Product[];
  sales: Sale[];
  expenses?: Expense[];
}

export interface RestoreResult {
  products: number;
  sales: number;
  expenses?: number;
}