# lib/api-zod/src/generated/types/backupSnapshot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabricGhar API specification — shared shop data scoped by user email
 * OpenAPI spec version: 0.1.0
 */
import type { Expense } from "./expense";
import type { Product } from "./product";
import type { Sale } from "./sale";

export interface BackupSnapshot {
  exportedAt?: string;
  products: Product[];
  sales: Sale[];
  expenses?: Expense[];
}