# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabricGhar API specification — shared shop data scoped by user email
 * OpenAPI spec version: 0.1.0
 */
import type { Sale } from "./sale";

export interface DashboardSummary {
  todayRevenue: number;
  todayProfit: number;
  totalRevenue: number;
  totalProfit: number;
  productCount: number;
  lowStockCount: number;
  todaySales: Sale[];
}