# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabricGhar API specification — shared shop data scoped by user email
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseInput {
  /** @minLength 1 */
  cat: string;
  /** @minimum 0 */
  amount: number;
  note?: string;
  /** Optional ISO date string (defaults to now) */
  date?: string;
}