# lib/api-zod/src/generated/types/monthlyReportEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabricGhar API specification — shared shop data scoped by user email
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyReportEntry {
  /** YYYY-MM */
  month: string;
  revenue: number;
  profit: number;
  count: number;
}