# lib/api-zod/src/generated/types/reportData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabricGhar API specification — shared shop data scoped by user email
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyReportEntry } from "./monthlyReportEntry";
import type { ProductReportEntry } from "./productReportEntry";

export interface ReportData {
  totalRevenue: number;
  totalCost: number;
  totalProfit: number;
  profitMargin: number;
  byMonth: MonthlyReportEntry[];
  topProducts: ProductReportEntry[];
}