# lib/api-zod/src/generated/types/saleInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FabricGhar API specification — shared shop data scoped by user email
 * OpenAPI spec version: 0.1.0
 */

export interface SaleInput {
  pid: string;
  /** @minimum 1 */
  qty: number;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  disc: number;
  note?: string;
}